backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Task queue
celery==5.4.0
redis==5.2.1
kombu==5.4.2

# LLM clients
anthropic==0.40.0
google-generativeai==0.8.3
openai==1.57.4

# HTTP / scraping
httpx==0.28.1
aiohttp==3.11.10
beautifulsoup4==4.12.3
lxml==5.3.0
duckduckgo-search==7.0.1

# Config
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Utils
tenacity==9.0.0
structlog==24.4.0
python-dateutil==2.9.0
pytz==2024.2
